backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
gunicorn>=21.0.0
python-multipart>=0.0.6
